requirements.txt
faiss-cpu              1.8.0.post1
grpcio-status          1.75.1
langchain-community    0.3.31
langchain-google-genai 2.1.12
langchain-huggingface  0.3.1
pip                    22.0.2
pypdf                  6.1.1
sentence-transformers  5.1.1
streamlit              1.50.0